add authentication and user profile management features with JWT support

+++ src/main/java/com/kakadev/cloudsharing/service/impl/UserServiceImpl.java
@@ -1,33 +1,52 @@
 package com.kakadev.cloudsharing.service.impl;
 
 import com.kakadev.cloudsharing.dto.request.CreateUserRequestDTO;
-import com.kakadev.cloudsharing.dto.request.UpdateUserRequestDTO;
+import com.kakadev.cloudsharing.dto.request.UpdateUserPasswordRequestDTO;
+import com.kakadev.cloudsharing.dto.request.UpdateUserProfileRequestDTO;
 import com.kakadev.cloudsharing.dto.response.UserResponseDTO;
 import com.kakadev.cloudsharing.exception.AppException;
 import com.kakadev.cloudsharing.exception.ErrorCode;
 import com.kakadev.cloudsharing.mapper.UserMapper;
+import com.kakadev.cloudsharing.model.entity.Role;
 import com.kakadev.cloudsharing.model.entity.User;
 import com.kakadev.cloudsharing.model.enums.AuthProvider;
+import com.kakadev.cloudsharing.repository.RoleRepository;
 import com.kakadev.cloudsharing.repository.UserRepository;
 import com.kakadev.cloudsharing.service.UserService;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@FieldDefaults(level = lombok.AccessLevel.PRIVATE, makeFinal = true)
 public class UserServiceImpl implements UserService {
 
-    private final UserRepository userRepository;
-    private final UserMapper userMapper;
+    UserRepository userRepository;
+    RoleRepository roleRepository;
+    UserMapper userMapper;
+    PasswordEncoder passwordEncoder;
 
     @Override
-    public UserResponseDTO createUser(CreateUserRequestDTO createUserRequestDTO) {
+    public UserResponseDTO createUser(
+            CreateUserRequestDTO createUserRequestDTO
+    ) {
         User newUser = userMapper.toEntity(createUserRequestDTO);
 
+        Set<Role> roles = new HashSet<>();
+        roleRepository.findById(com.kakadev.cloudsharing.model.enums.Role.USER.getName())
+                .ifPresent(roles::add);
+
+        newUser.setRoles(roles);
         newUser.setProvider(AuthProvider.LOCAL);
+        newUser.setPassword(passwordEncoder.encode(newUser.getPassword()));
         newUser.setCredits(5);
 
         User savedUser = userRepository.save(newUser);
@@ -35,7 +54,9 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserResponseDTO getUserById(UUID userId) {
+    public UserResponseDTO getUserById(
+            UUID userId
+    ) {
         return userMapper.toResponseDTO(userRepository.findById(userId)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND)));
     }
@@ -50,12 +71,18 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserResponseDTO getMyProfile() {
-        // TODO: Implement authentication and get the current user
-        return null;
+        var context = SecurityContextHolder.getContext();
+        String email = context.getAuthentication().getName();
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+        return userMapper.toResponseDTO(user);
     }
 
     @Override
-    public UserResponseDTO updateUser(UUID userId, UpdateUserRequestDTO updateUserRequestDTO) {
+    public UserResponseDTO updateUserProfile(
+            UUID userId, UpdateUserProfileRequestDTO updateUserRequestDTO
+    ) {
         User existingUser = userRepository.findById(userId)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
 
@@ -65,7 +92,34 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    public UserResponseDTO updateUserPassword(
+            UUID userId, UpdateUserPasswordRequestDTO updateUserPasswordRequestDTO
+    ) {
+        User existingUser = userRepository.findById(userId)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+
+        if (!passwordEncoder.matches(
+                updateUserPasswordRequestDTO.getCurrentPassword(), existingUser.getPassword()
+        )) {
+            throw new AppException(ErrorCode.PASSWORD_INCORRECT);
+        }
+
+        if (!updateUserPasswordRequestDTO.getNewPassword().equals(
+                updateUserPasswordRequestDTO.getConfirmPassword()
+        )) {
+            throw new AppException(ErrorCode.PASSWORD_MISMATCH);
+        }
+
+        existingUser.setPassword(passwordEncoder.encode(updateUserPasswordRequestDTO.getNewPassword()));
+        User updatedUser = userRepository.save(existingUser);
+        return userMapper.toResponseDTO(updatedUser);
+    }
+
+    @Override
     public void deleteUser(UUID userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.deleteById(userId);
     }
 }
